Rename `livekit_server` to `livekit_api` as it is not a server
