+++ crates/collab/Cargo.toml
@@ -40,7 +40,7 @@ google_ai.workspace = true
 hex.workspace = true
 http_client.workspace = true
 jsonwebtoken.workspace = true
-livekit_server.workspace = true
+livekit_api.workspace = true
 log.workspace = true
 nanoid.workspace = true
 open_ai.workspace = true
