--- crates/livekit_server/src/livekit_server.rs
@@ -1,3 +1,0 @@
-pub mod api;
-pub mod proto;
-pub mod token;
